refactor(dao): 修改接口的入参类型，使匹配GORM的ID类型

+++ dao/memory/post_dao_impl.go
@@ -23,6 +23,6 @@ func NewPostDaoImplInstance() *PostDaoImpl {
 }
 
 // 实现Topic_dao的方法
-func (*PostDaoImpl) QueryPostListByTopicId(id uint64) ([]*entity.Post, error) {
+func (*PostDaoImpl) QueryPostListByTopicId(id uint) ([]*entity.Post, error) {
 	return repository.PostIndexMap[id], nil
 }
